Feat: backward pagerank

+++ Connection.cpp
@@ -1,8 +1,8 @@
 #include "Connection.h"
 
-Connection::Connection(int id, int origin, int destination, double weight)
+Connection::Connection(int id, int origin, int destination, float weight)
     : id(id), origin(origin), destination(destination), weight(weight) {}
 
-void Connection::actualizarPeso(double nuevoPeso) {
+void Connection::actualizarPeso(float nuevoPeso) {
     weight = nuevoPeso;
 }
