+++ Node.cpp
@@ -1,6 +1,6 @@
 #include "Node.h"
 
-void Node::actualizarRelevancia(double nuevaRelevancia) {
+void Node::actualizarRelevancia(float nuevaRelevancia) {
     relevanciaPrev = relevanciaActual;
     relevanciaActual = nuevaRelevancia;
 }
